Revert "Dynamic Player/Rival Avatars"
This reverts commit 22f75f2d380752579b9b3a7bfbb229b9a7a42b9b.

+++ src/frontier_util.c
@@ -52,7 +52,7 @@ struct FrontierBrainMon
 struct FrontierBrain
 {
     u16 trainerId;
-    u16 objEventGfx;
+    u8 objEventGfx;
     u8 isFemale;
     const u8 *lostTexts[2];
     const u8 *wonTexts[2];
